@@ -1,6 +1,15 @@
+<script lang="ts" module>
+    import type { Snippet } from "svelte";
+    
+    interface Props {
+        id: string, 
+        type: string, 
+        children: Snippet
+    }
+</script>
+
 <script lang="ts">
-    export let id = "";
-    export let type: string;
+    let { id = '', type = $bindable<string>(), children }: Props = $props();
 
     function set(e: MouseEvent) {
         const { id } = e.currentTarget as HTMLButtonElement;
@@ -8,8 +17,8 @@
     }
 </script>
 
-<button {id} class="clear outline" on:click={set}>
-    <slot />
+<button {id} class="clear outline" onclick={set}>
+    {@render children?.()}
 </button>
 
 <style>
