@@ -1,7 +1,8 @@
 import { name, repository } from 'package.json'
+import { mount } from 'svelte';
 import App from './App.svelte';
 
-export default new App({
+export default mount(App, {
     target: document.body,
     props: { name }
 });
